move operator2Proto to CometExecRule
